Mwbot: Use Axios instance
Introduce an Axios instance as a protected property to manage constant default request options.
Since HTTP agents don't need to (and shouldn't) be deep-cloned, they are no longer included in `defaultRequestOptions`.

Also implement a re-injection mechanism for `params` into `requestOptions`:
`_request` removes the `params` property for POST requests to avoid MediaWiki API errors, so it must be restored when a retry is triggered.