fix: prevent truncated responses with `read()` (`rvlimit=1`)
Set `rvlimit` to 1, since there is no need to reference the contents of
old revisions. This change helps reduce the likelihood of encountering
the error "This result was truncated because it would otherwise be larger
than the limit of 12,582,912 bytes". However, this is only a partial
mitigation; a complete solution still needs to be implemented.